Unite all Composer tests into one file
