--- cypress/integration/composer/createAndDeleteArticle.ts
@@ -1,12 +1,0 @@
-import { fetchAndSetCookie } from '../../utils/networking';
-import { checkVars } from '../../utils/vars';
-import { inATemporaryArticle } from '../../utils/composer/inATemporaryArticle';
-
-describe('Composer Basic Behaviour Tests', () => {
-  beforeEach(() => {
-    checkVars();
-    fetchAndSetCookie({ visitDomain: true });
-  });
-
-  inATemporaryArticle('Do nothing but create and delete');
-});
